requirements.txt
apscheduler
google-api-python-client
google-auth
google-auth-oauthlib
icalendar
caldav
SQLAlchemy
python-dotenv
tenacity
